requirements.txt
supabase==0.7.1
disnake==2.7
python-dotenv==0.20